+++ Source/Bogus.Tests/ExtensionTests/ExtensionTest.cs
@@ -4,7 +4,7 @@ using Bogus.Extensions.Italy;
 using Bogus.Extensions.Portugal;
 using FluentAssertions;
 using System;
-using Bogus.Extensions.Romanian;
+using Bogus.Extensions.Romania;
 using Xunit;
 
 namespace Bogus.Tests.ExtensionTests;
